Fix bounds for Date plotting in years
When plotting Dates in years, Gadfly previously chose x_max as the year of the maximum date.  This would truncate any data points that are in the most recent year after January 1.

This patch changes it to be the year of the maximum date plus a year less a day, retaining the current behavior for dates that fall exactly on January one, and choosing better bounds for all other dates.