Hide API pages in the sidebar
Use the new `hide` feature of Documenter to hide a bunch of pages in
the sidebar. Also, move the index pages to the Library section since
they feel more like an API reference rather than a manual page.

Also, the enums for the format option are now deprecated, so switch to
`format = :html` instead.