try to be less clever and add ability to stack densities
My original implementation was too clever in that it figured out the
orientation of density and violin plots automatically. The logic ended
up being quite convoluted and so I switched back to using the standard
`orientation` logic and an explicit flag for whether a density plot is a
violin or not.

This commit also adds the ability to stack either raw densities and to
create conditional density distributions.